==> ara_code/0d_compare_raw_data_correction.ipynb <==
## Comparison of Raw Data Correction Methods
Methods compared:
1. Brianna's linear model (linb) (only single mutants and wild type grown on one flat [not four flats])

    For sets with many flats:
        
        formula = f'{col_name} ~ Genotype + (1|Column) + (1|Row) + (1|Flat)'
    
    For sets with one flat:
        
        formula = f'{col_name} ~ Genotype + (1|Column) + (1|Row)'

2. Estimation of marginal means for each genotype using lmer in R (single, double, and wild type)

    Per set, per flat:
    
        formula = TSC ~ Subline + (1|Column) + (1|Row)

3. Spatial Analysis with SpATS (single, double, and wild type)

import datatable as dt
import pandas as pd

### Read in the corrected raw datasets

# Results on single mutants (that were grown on only one flat) for the lmer model Brianna ran in python
og_bri = dt.fread('../ara_data/brianna_comparemean_tolmer_df_withrelative.csv').to_pandas()
og_bri.head()

og_bri.loc[og_bri['Set'].str.contains('1'),:] # 1 is a set grown on four flats; these sets were excluded in her analysis

# Results on single and double mutants for the lmer model I ran in R (should be emulating Brianna's results)
linb = dt.fread('../ara_data/double_mutant_fitness_data_05312024_all_corrected_linear_b.txt').to_pandas()
# linb = linb.loc[:,~linb.columns.str.contains('.y')]
# linb.insert(5, 'Genotype', linb.Subline.str.split('-').str[1])
# linb.to_csv('../ara_data/double_mutant_fitness_data_05312024_all_corrected_linear_b.txt', sep='\t', index=False)
linb.head()

# Results on single and double mutants for the lmer model ran per set per flat
lin = dt.fread('../ara_data/double_mutant_fitness_data_05312024_all_corrected_linear.txt').to_pandas()
# lin = lin.loc[:,~lin.columns.str.contains('.y')]
# lin.insert(5, 'Genotype', lin.Subline.str.split('-').str[1])
# lin.to_csv('../ara_data/double_mutant_fitness_data_05312024_all_corrected_linear.txt', sep='\t', index=False)
lin.head()

# Results on single and double mutants for the spatial analysis model ran per set per flat
spa = dt.fread('../ara_data/double_mutant_fitness_data_05312024_TSC_corrected_SpATS.txt').to_pandas()
spa.head()

spa[['Set', 'Flat', 'Genotype', 'TSC', 'fit.TSC$fitted']].groupby(['Set', 'Flat', 'Genotype']).mean()
# For some reason, the spatial model predicts the mean of the genotypes perfectly,
# but if you look at the mean of the sublines, it's not perfect.

spa[['Set', 'Flat', 'Subline', 'TSC', 'fit.TSC$fitted']].groupby(['Set', 'Flat', 'Subline']).mean()

og_bri.shape, linb.shape, lin.shape, spa.shape

# Reshape Brianna's data
og_Bri = og_bri[['Set', 'WT_fitlmer', 'MA_fitlmer', 'MB_fitlmer']].melt(id_vars = 'Set', value_name='TSC_corrected', var_name='Genotype')
og_Bri.Genotype = og_Bri.Genotype.str.split('_').str.get(0)
og_Bri.head()

bri_raw = og_bri[['Set', 'WT_avg', 'MA_avg', 'MB_avg']].melt(id_vars = 'Set', value_name='TSC_avg_raw', var_name='Genotype')
bri_raw.Genotype = bri_raw.Genotype.str.split('_').str.get(0)
bri_raw.head()

### Combine the datasets and calculate the correlation between genotype averages per set for Total Seed Count

# Merge corrected values with the mean of the raw data for single mutants
# Averages are not counted with flat considered since Brianna added flat as a random effect in her analysis
corrected = pd.merge(linb[['Set', 'Genotype', 'TSC_emmean']].\
    groupby(['Set', 'Genotype']).mean(), lin[['Set', 'Genotype', 'TSC_emmean']].\
    groupby(['Set', 'Genotype']).mean(), left_index=True, right_index=True, how='left') # Linear model results to compare with Brianna's

corrected = pd.merge(corrected, spa[['Set', 'Genotype', 'fit.TSC$fitted']].\
    groupby(['Set', 'Genotype']).mean(), 
    left_on=['Set', 'Genotype'], right_index=True, how='left') # Spatial analysis results

corrected = pd.merge(corrected, og_Bri, left_index=True,
    right_on=['Set', 'Genotype'], how='left') # Brianna's python results

corrected = pd.merge(bri_raw, corrected, on=['Set', 'Genotype'], how='left') # Brianna's raw mean data

corrected = pd.merge(corrected, lin[['Set', 'Genotype', 'TSC']].groupby(['Set', 'Genotype']).mean(),
    left_on=['Set', 'Genotype'], right_index=True, how='left') # Raw mean data (to compare with Brianna's)

corrected.columns = ['Set', 'Genotype', 'TSC_avg_raw_bri', 'Brianna_rerun', 'Linear',
                     'SpATS', 'Brianna_og', 'TSC_raw_avg']

corrected

corrected.select_dtypes('float').corr(method='pearson')

# Since Brianna only did the single mutant data, there are NAs when I combine her results with mine
corrected.select_dtypes('float').dropna().corr(method='pearson')

corrected.loc[corrected['Set'].str.contains('845'),:]

# Now include "Flat", since I built models per set per flat
corrected = pd.merge(linb[['Set', 'Flat', 'Genotype', 'TSC_emmean']].\
    groupby(['Set', 'Flat', 'Genotype']).mean(),
    lin[['Set', 'Flat', 'Genotype', 'TSC_emmean']].\
    groupby(['Set', 'Flat', 'Genotype']).mean(), 
    left_index=True, right_index=True, how='left') # Results from linear models

corrected = pd.merge(corrected, spa[['Set', 'Flat', 'Genotype', 'fit.TSC$fitted']].\
    groupby(['Set', 'Flat', 'Genotype']).mean(), 
    left_on=['Set', 'Flat', 'Genotype'], right_index=True, how='left') # Spatial analysis results

corrected = pd.merge(corrected, lin[['Set', 'Flat', 'Genotype', 'TSC']].\
    groupby(['Set', 'Flat', 'Genotype']).mean(),
    left_on=['Set', 'Flat', 'Genotype'], right_index=True, how='left') # Raw mean data

corrected.columns = ['Brianna_rerun', 'Linear', 'SpATS', 'TSC_raw_avg']

corrected

corrected.corr(method='pearson')

Conclusion:

Of the three approaches I used, it seems to not matter which one I go with. 
But, I don't trust the spatial model, so I won't use that one. Instead, I will 
use the "Linear" model built per set per flat, instead of Brianna's re-run, where 
she built a linear model per set and flat was a random variable for those sets 
grown on 4 flats.

Remaining mystery:

I believe my analysis is correct, so I don't know why there is disagreement with 
what Brianna built in python. I made sure to keep the WT genotype as the reference 
level for the three approaches. And the models I ran agree with the average total 
seed count value, which also agrees with Brianna's averages if you see the dataframe, 
but I don't know why the correlation is so low. The correlation of her results 
is only ~.46 with "Brianna_rerun", "Linear", "SpATS", and "TSC_raw_avg".
Brianna said she has to re-run her stuff and clean up her code, so in the mean time, 
I think I will move forward.

Mystery solved:

Brianna used an older version of the dataset, hence why there is disagreement.
She re-ran the model with the updated raw data (that I used) and at least for 
set 845, our numbers agree, so the correlation should go up to .99 if I were to 
include here.

### Compare the averages for the rest of the fitness traits

all_corrected = pd.concat([lin[['Set', 'Flat', 'Genotype', 'GN', 'PG', 'DTB',
       'LN', 'DTF', 'SN', 'WO', 'FN', 'SPF', 'TSC', 'SH']], 
       lin.loc[:,lin.columns.str.endswith('_emmean')]], axis=1)
all_corrected2 = pd.concat([linb[['Set', 'Flat', 'Genotype']], 
       linb.loc[:,linb.columns.str.endswith('_emmean')]], axis=1)
all_corrected = pd.merge(all_corrected, all_corrected2, on=['Set', 'Flat', 'Genotype'], how='left', suffixes=['_lin', '_linb'])
all_corrected = all_corrected.groupby(['Set', 'Flat', 'Genotype']).mean()
all_corrected

correlations = all_corrected.corr(method='pearson')
correlations

Conclusion:

Linb (flat included as random effect for sets grown on multiple flats) yields 
genotype averages that are similar (PCC > 0.95) to the raw data averages for all 
the traits. 

Lin (per set per flat) yields genotype averages that are similar only to total 
seed count raw data averages (PCC > 0.99). With LN (PCC = 0.30), with SN 
(PCC = 0.45), with SPF (PCC = 0.50), and with SH (PCC = 0.47). The other traits 
have -0.23 < PCC < 0.09.
