==> ara_code/1a_make_regression_feature_tables.ipynb <==
# Create feature tables for regression
Use corrected single mutant relative fitness and genetic interaction neutrality functions (Mani 2008 PNAS) to predict the relative fitness of the double mutant

- 9 features:

  Wa (max)

  Wb (min)

  additive (Wa + Wb)

  difference (Wa - Wb)

  multiplicative (Wa * Wb)
  
  log2_mani (log2[((2^Wa) − 1)*((2^Wb) − 1) + 1])

  pslog10_additive (pseudolog10(Wa * Wb))

  pslog10_difference (pseudolog10(Wa / Wb))

  mean ((Wa + Wb)/2)

Labels:
- Wab as total seed count (TSC)

Relative fitness (W): total seed count (corrected for batch effects)

Wa: relative fitness of gene mutant A

Wb: relative fitness of gene mutant B

Wab: relative fitness of the double mutant

# Read in the data
import pandas as pd
# data = pd.read_csv('../ara_data/double_mutant_fitness_data_05312024_all_corrected_linear_b.txt', sep='\t')
data = pd.read_csv('../ara_data/fitness_data_for_Kenia_09232024_all_corrected_brianna.txt', sep='\t')
data

## Calculate the relative fitness of the single and double mutants using the corrected trait values

# Subset the corrected trait values
data_avg = data.loc[:,data.columns.isin(['Set', 'Flat', 'Genotype'])\
    | (data.columns.str.endswith('_emmean'))].\
    groupby(['Set', 'Flat', 'Genotype']).mean()
data_avg

# Calculate the relative fitness
W = data_avg.index.to_frame()
for trait in data_avg.columns:
    tmp = data_avg[trait].reset_index().pivot(index=['Set', 'Flat'], columns='Genotype', values=trait) # reshape tmp to wide
    W_tmp = tmp.apply(lambda x: x / x['WT'], axis=1) # relative fitness
    W_tmp = W_tmp.reset_index().melt(id_vars=['Set', 'Flat'], var_name='Genotype', value_name=f'W_{trait}') # reshape W_tmp back to long
    W_tmp.set_index(['Set', 'Flat', 'Genotype'], inplace=True)
    W = pd.concat([W, W_tmp], ignore_index=False, axis=1)
W

print(data_avg.index.get_level_values('Set').nunique())
print(W.index.get_level_values('Set').nunique())

# Check which sets do not have double mutant information
for name, group in W.groupby(level='Set'):
    geno_levels = group.dropna(axis=0, how='all').index.get_level_values('Genotype').unique()
    if 'DM' not in geno_levels.values:
        print(geno_levels)
        print(f'Drop set {name}')

# This was done for 20240531 data. There were 137 sets, and 2 were dropped, leaving 135 sets.
# Drop sets that do not have double mutant information
# W = W.loc[~W.index.get_level_values('Set').isin(['845', '845E']),:]

W.drop(columns=['Set', 'Flat', 'Genotype'], inplace=True)
W

## Feature tables for total seed count (TSC) and the other traits
1. Ensure the first single mutant column has the greater value than the other single mutant
2. Calculate the 6 neutrality functions using the SMF data

def pseudolog10(x):
  # Pseudo-Logarithm of base 10, is defined for all real numbers. Use instead of
  # log10, which returns infinite/NaN values for x <= 0
  return (np.log((x/2) + np.sqrt((x/2)**2 + 1)) / np.log(10))

import numpy as np
for trait in W.columns.values:
    print(trait)
    trait_df = W[trait].unstack()
    trait_df.drop(columns='WT', inplace=True) # drop wild type
    trait_df.index = trait_df.index.get_level_values('Set').astype(str) # drop flat, since values don't change for sets with flat included as the random effect
    trait_df.index = trait_df.index.get_level_values('Set').str.strip()
    trait_df = trait_df.groupby('Set').mean()
    trait_df.dropna(axis=0, how='all', inplace=True) # drop rows with all NaN values (this particular trait was not collected for these sets)

    # reorder the values in MA and MB so that MA is always greater than MB
    trait_df['MA_new'] = trait_df['MA'].where(trait_df['MA'] > trait_df['MB'], trait_df['MB'])
    trait_df['MB_new'] = trait_df['MB'].where(trait_df['MA'] > trait_df['MB'], trait_df['MA'])

    # calculate the expected double mutant relative fitness using the neutrality functions
    trait_df['mean'] = (trait_df['MA_new'] + trait_df['MB_new']) / 2
    trait_df['multiplicative'] = trait_df['MA_new'] * trait_df['MB_new']
    trait_df['additive'] = trait_df['MA_new'] + trait_df['MB_new'] - 1
    trait_df['difference'] = trait_df['MA_new'] - trait_df['MB_new']
    trait_df['log2_mani'] = np.log2(((2**trait_df['MA_new']) - 1) * ((2**trait_df['MB_new']) - 1) + 1)
    trait_df['pslog10_additive'] = trait_df.apply(lambda x: pseudolog10(x['MA_new'] * x['MB_new']), axis=1)
    trait_df['pslog10_difference'] = trait_df.apply(lambda x: pseudolog10(x['MA_new'] / x['MB_new']), axis=1)
    # pseudolog10(trait_df['MA_new'] * trait_df['MB_new'])
    # trait_df['log10_difference'] = pseudolog10(trait_df['MA_new'] / trait_df['MB_new'])

    # create sample IDs
    ID = 'Set_' + trait_df.index.astype(str)
    trait_df.insert(0, 'ID', ID)
    trait_df.set_index('ID', inplace=True)

    # save the feature table for the regression models to predict the trait using GI defs
    trait_df.rename(columns={'DM': trait}, inplace=True)
    trait_df.replace([np.inf, -np.inf], np.nan) # replace inf with NaN
    trait_df.dropna(axis=0, how='all', inplace=True) # drop rows with all NaN values
    trait_df.loc[:,[trait, 'MA_new', 'MB_new', 'mean', 'multiplicative', 
        'additive', 'difference', 'log2_mani', 'pslog10_additive', 'pslog10_difference']].\
        to_csv(f'../ara_data/1_feature_tables/{trait}_feature_table.tsv', sep='\t')
    trait_df.corr(method='pearson').to_csv(f'../ara_data/1_feature_tables/{trait}_feature_correlation.tsv', sep='\t')
    print(trait_df.shape)

{trait}_feature_table.tsv and {trait}_feature_correlation.tsv were saved to ara_data/1_feature_tables/20240923_melissa_ara_data_features

## Stratified K-fold train-test split

from sklearn.model_selection import StratifiedKFold
# sklearn v1.2.2
np.random.seed(20240606)

for trait in W.columns.values:
    trait_df = pd.read_csv(
        f'../ara_data/1_feature_tables/20240923_melissa_ara_data_features/{trait}_feature_table.tsv',
        sep='\t', index_col=0)
    print(trait, 'trait_df', trait_df.shape)

    # Create bins
    trait_df['label_bin'] = pd.cut(trait_df[trait], bins=[-np.inf,
        np.quantile(trait_df[trait], 0.25), np.quantile(trait_df[trait], 0.5),
        np.quantile(trait_df[trait], 0.75), np.inf], labels=[0, 1, 3, 4])

    # Apply stratified k-fold train-test split
    y = trait_df['label_bin']
    X = trait_df.drop(columns=[trait, 'label_bin'], axis=1)
    skf = StratifiedKFold(n_splits=6, random_state=20240606, shuffle=True)

    i = 0
    for train_idx, test_idx, in skf.split(X, y):
        if i == 0:
            # Write test set to file
            with open(f'../ara_data/{trait}_test_instances.txt', 'w') as f:
                for ID in test_idx:
                    f.write("%s\n" % trait_df.iloc[ID,:].name)
        else:
            break
        i += 1

{trait}_test_instsances.txt files were saved to ara_data/1_feature_tables/20240923_melissa_ara_data_features

## Run XGBoost Regression model on TSC (total seed count)

### Use shap conda environment and run on the command line
__Using all 9 neutrality functions:__
```python
python 1b_xgb_regression.py \
    -X ../ara_data/1_feature_tables/20240923_melissa_ara_data_features/W_TSC_emmean_feature_table.tsv \
    -y_name W_TSC_emmean \
    -test ../ara_data/1_feature_tables/20240923_melissa_ara_data_features/W_TSC_emmean_test_instances.txt \
    -save ../output/1_xgb_regression_ara/20240923_melissa_ara_data_results \
    -prefix W_TSC_emmean \
    -tag use_Neutrality_Funcs \
    -fold 5 -n 20 -plot t
```

__Using only the four established neutrality functions:__
```python
python 1b_xgb_regression.py \
    -X ../ara_data/1_feature_tables/20240923_melissa_ara_data_features/W_TSC_emmean_feature_table.tsv \
    -y_name W_TSC_emmean \
    -feat_list MB_new,multiplicative,additive,log2_mani \
    -test ../ara_data/1_feature_tables/20240923_melissa_ara_data_features/W_TSC_emmean_test_instances.txt \
    -save ../output/1_xgb_regression_ara/20240923_melissa_ara_data_results \
    -prefix W_TSC_emmean_established_neut_funcs \
    -tag use_Established_Neutrality_Funcs \
    -fold 5 -n 20 -plot t
```

__Using only the single mutant information:__
```python
python 1b_xgb_regression.py \
    -X ../ara_data/1_feature_tables/20240923_melissa_ara_data_features/W_TSC_emmean_feature_table.tsv \
    -y_name W_TSC_emmean \
    -feat_list MA_new,MB_new \
    -test ../ara_data/1_feature_tables/20240923_melissa_ara_data_features/W_TSC_emmean_test_instances.txt \
    -save ../output/1_xgb_regression_ara/20240923_melissa_ara_data_results \
    -prefix W_TSC_emmean_SMF_only \
    -tag use_SMF_only \
    -fold 5 -n 20 -plot t
```

### Gradient Boosting Model (Use ml-pipe1.5 conda Environment)
__Using all 9 neutrality functions:__
```python
python 1b_gb_regression.py \
    -X ../ara_data/W_TSC_emmean_feature_table.txt \
    -y_name W_TSC_emmean \
    -test ../ara_data/W_TSC_emmean_test_instances.txt \
    -save ../output/1_gb_regression_ara/ \
    -prefix W_TSC_emmean_gb \
    -tag use_SMF_and_Neutrality_Funcs \
    -fold 5 -n 10 -plot t
```

__Using only the four established neutrality functions:__
```python
```

__Using only the single mutant information:__
```python
```

### Multi-layer Perceptron
__Using all 9 neutrality functions:__


__Using only the four established neutrality functions:__

## Feature tables for predicting co-function from fitness data only

## Feature tables for predicting co-function from the GI definitions
