==> ara_code/2a_feature_importance.ipynb <==
### Feature importance of the models:
1. Using all neutrality functions
2. Using only the established neutrality functions

### Scenario 1. Using all neutrality functions

# Old run: I calculated log_mani definition wrong, so this model is not valid.
# Model results are in folder Yeast-Ara-GI/output/1_xgb_regression_ara/20240531_melissa_ara_data_results
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns

imp = pd.read_csv("../output/1_xgb_regression_ara/W_TSC_emmean_imp.csv", index_col=0)
imp["mean_imp"] = imp.mean(axis=1)
imp["std_imp"] = imp.std(axis=1)
imp = imp.sort_values("mean_imp", ascending=False)

# plot mean_imp as a barplot with error bars and sort by mean_imp
plt.figure(figsize=(10, 5))
sns.barplot(x=imp.index, y="mean_imp", data=imp, yerr=imp["std_imp"])
plt.xticks(rotation=90)
plt.ylabel("Mean Importance")
plt.title("Feature Importance")
plt.tight_layout()
plt.savefig("../output/2_feature_importance_ara/all_neut_funcs_feature_importance.pdf")
plt.close()

# Oct. 15, 2024 model -- this is the correct one.
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns

imp = pd.read_csv("../output/1_xgb_regression_ara/20240923_melissa_ara_data_results/W_TSC_emmean_imp.csv", index_col=0)
imp["mean_imp"] = imp.mean(axis=1)
imp["std_imp"] = imp.std(axis=1)
imp = imp.sort_values("mean_imp", ascending=False)

# plot mean_imp as a barplot with error bars and sort by mean_imp
plt.figure(figsize=(10, 5))
sns.barplot(x=imp.index, y="mean_imp", data=imp, yerr=imp["std_imp"])
plt.xticks(rotation=90)
plt.ylabel("Mean Importance")
plt.title("Feature Importance")
plt.tight_layout()
plt.savefig("../output/2_feature_importance_ara/20240923_melissa_ara_data_figs/all_neut_funcs_feature_importance.pdf")
plt.close()

### Scenario 2. Using only the established neutrality functions

# Old run: I calculated log_mani definition wrong, so this model is not valid.
# Model results are in folder Yeast-Ara-GI/output/1_xgb_regression_ara/20240531_melissa_ara_data_results
imp = pd.read_csv("../output/1_xgb_regression_ara/W_TSC_emmean_established_neut_funcs_imp.csv", index_col=0)
imp["mean_imp"] = imp.mean(axis=1)
imp["std_imp"] = imp.std(axis=1)
imp = imp.sort_values("mean_imp", ascending=False)

# plot mean_imp as a barplot with error bars and sort by mean_imp
plt.figure(figsize=(10, 5))
sns.barplot(x=imp.index, y="mean_imp", data=imp, yerr=imp["std_imp"])
plt.xticks(rotation=90)
plt.ylabel("Mean Importance")
plt.title("Feature Importance")
plt.tight_layout()
plt.savefig("../output/2_feature_importance_ara/established_neut_funcs_feature_importance.pdf")
plt.close()

# Oct. 15, 2024 model -- this is the correct one.
imp = pd.read_csv("../output/1_xgb_regression_ara/20240923_melissa_ara_data_results/W_TSC_emmean_established_neut_funcs_imp.csv", index_col=0)
imp["mean_imp"] = imp.mean(axis=1)
imp["std_imp"] = imp.std(axis=1)
imp = imp.sort_values("mean_imp", ascending=False)

# plot mean_imp as a barplot with error bars and sort by mean_imp
plt.figure(figsize=(10, 5))
sns.barplot(x=imp.index, y="mean_imp", data=imp, yerr=imp["std_imp"])
plt.xticks(rotation=90)
plt.ylabel("Mean Importance")
plt.title("Feature Importance")
plt.tight_layout()
plt.savefig("../output/2_feature_importance_ara/20240923_melissa_ara_data_figs/established_neut_funcs_feature_importance.pdf")
plt.close()
